creacion de servicios de usuario, rol y tipo de identificacion

+++ coworking/coworking/src/main/java/com/backend/digitalhouse/coworking/dto/entrada/calificacion/CalificacionEntradaDto.java
@@ -8,17 +8,22 @@ import javax.validation.constraints.Pattern;
 public class CalificacionEntradaDto {
     @NotNull(message = "Este campo no puede ser nulo")
     @Pattern(regexp = "^[1-5]$", message = "La calificacion debe tener un numero de 1 a 5")
-    private long puntuacion;
+    private int puntuacion;
 
     public CalificacionEntradaDto() {
     }
 
-    public CalificacionEntradaDto(long puntuacion) {
+    public CalificacionEntradaDto(int puntuacion) {
         this.puntuacion = puntuacion;
     }
 
-    public long getPuntuacion() {return puntuacion;}
-    public void setPuntuacion(long puntuacion) {this.puntuacion = puntuacion;}
+    public int getPuntuacion() {
+        return puntuacion;
+    }
+
+    public void setPuntuacion(int puntuacion) {
+        this.puntuacion = puntuacion;
+    }
 
     @Override
     public String toString() {
